feat: rm get lib logs method
